+++ src/cosmos/info/thorchainruneInfo.ts
@@ -28,10 +28,27 @@ const networkInfo: CosmosNetworkInfo = {
     url: 'https://rpc.ninerealms.com',
     headers: { 'x-client-id': '{{ninerealmsClientId}}' }
   },
-  archiveNode: {
-    url: 'https://rpc-v1.ninerealms.com',
-    headers: { 'x-client-id': '{{ninerealmsClientId}}' }
-  }
+  archiveNodes: [
+    {
+      blockTimeRangeSeconds: {
+        start: 1647912564649 // 2022-03-22T01:29:24.649Z
+        // end: TBD
+      },
+      endpoint: {
+        url: 'https://rpc-v1.ninerealms.com',
+        headers: { 'x-client-id': '{{ninerealmsClientId}}' }
+      }
+    }
+    // {
+    //   blockTimeRangeSeconds: {
+    //     start: TBD
+    //   },
+    //   endpoint: {
+    //     url: 'https://rpc-v2.ninerealms.com',
+    //     headers: { 'x-client-id': '{{ninerealmsClientId}}' }
+    //   }
+    // }
+  ]
 }
 
 const currencyInfo: EdgeCurrencyInfo = {